refactor: move cursor.hpp to Cursor.hpp
- Move the file src/Editor/cursor/cursor.hpp to src/Cursor/Cursor.hpp
