+++ src/CMakeLists.txt
@@ -13,7 +13,7 @@ target_include_directories(kilo
 target_sources(kilo
     PRIVATE
         "${PROJECT_SOURCE_DIR}/src/Constants/Constants.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Editor/cursor/cursor.hpp"
+        "${PROJECT_SOURCE_DIR}/src/Cursor/Cursor.hpp"
 
         "${PROJECT_SOURCE_DIR}/src/Editor/Editor.hpp"
         "${PROJECT_SOURCE_DIR}/src/Editor/Editor.cpp"
