+++ src/Editor/Editor.cpp
@@ -24,10 +24,10 @@
 #include "Editor.hpp"
 
 #include "Constants/Constants.hpp"
+#include "Cursor/Cursor.hpp"
 #include "Terminal/File.hpp"
 #include "Terminal/window/window.hpp"
 #include "Utilities/Utilities.hpp"
-#include "cursor/cursor.hpp"
 #include "offset/offset.hpp"
 #include "screen_buffer/screen_buffer.hpp"
 #include <algorithm>
