+++ src/Editor/config/config.hpp
@@ -1,7 +1,7 @@
 #ifndef EDITOR_CONFIG_HPP
 #define EDITOR_CONFIG_HPP
 
-#include "Editor/cursor/cursor.hpp"
+#include "Cursor/Cursor.hpp"
 #include "Editor/offset/offset.hpp"
 #include "Editor/screen_buffer/screen_buffer.hpp"
 #include "Terminal/mode/terminal_mode.hpp"
