+++ tests/CMakeLists.txt
@@ -21,7 +21,7 @@ target_include_directories(tests
 
 target_sources(tests
     PRIVATE
-        "${PROJECT_SOURCE_DIR}/src/Editor/cursor/cursor.hpp"
+        "${PROJECT_SOURCE_DIR}/src/Cursor/Cursor.hpp"
         Cursor.test.cpp
 
         "${PROJECT_SOURCE_DIR}/src/Editor/Editor.hpp"
